fix some window 64bit warnings

+++ smtk/Qt/qtComboItem.cxx
@@ -127,7 +127,7 @@ void qtComboItem::updateItemData()
   QList<QString> discreteVals;
   for (i = 0; i < itemDef->numberOfDiscreteValues(); i++)
     {
-    discreteVals.push_back(itemDef->discreteEnum(i).c_str());
+    discreteVals.push_back(itemDef->discreteEnum(static_cast<int>(i)).c_str());
     }
 
   foreach(QComboBox* combo, this->Internals->comboBoxes)
